always create imagemanifest at root of project and set to 'content' item type
imagemanifest discovery requires the file to be located at the root of project (so don't give the user the option to save at another location). If the user wants a different filename they can alter it after the file has been generated.

'content' item type is required for the `buildPropertyStorage.SetItemAttribute(itemId, "IncludeInVSIX", "true")` call to succeed (and therefore necessary for imagemanifest discovery)